WIP: Remove non-type-safe built-in op accessors
The Object... varargs methods make it more difficult for the programmer
to verify that the method they intend to call is actually the one which
_will_ be called, since the compiler falls back to the varargs signature
when the types do not fully match.

The major downside to removing these methods is that it makes some
scripts uglier: any time the types do not match a built-in method, but
would be converted on the fly to produce a successful match, the
programmer now needs to write e.g.:

    c = ops.run("math.add", a, b)

instead of:

    c = ops.math().add(a, b)

It is only five characters longer, but it is less intuitive.

That said, it is actually also possible to write:

    c = ops.run("add", a, b)

without the namespace, which is the same number of characters as the
ops.math().add(...) invocation above, and the operation will match all
namespaces, ultimately matching the same op in the vast majority of
cases.